requirements.txt
flask
line-bot-sdk
google-generativeai
requests
beautifulsoup4
python-dotenv
gspread
oauth2client
PyMuPDF
pandas
openpyxl
